:bug: Make the props object check stronger
With the current implementation an array with a single object would return "[object Object]". Using a ticky we can filter that out and get the proper type "[object Array]". Additionally we check if the props object has a "props" key, if it has, then it a VNode and therefore should be inserted as a children, not a props